Header e sidebar finalizados

+++ front-end/airbabeach/src/Components/Header.jsx
@@ -1,37 +1,69 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { Sidebar } from "./Sidebar";
 import { X } from 'phosphor-react'
 import "./Header.scss";
 import { useAuth } from "../contexts/auth";
+import Swal from 'sweetalert2'
+
 
 export function Header() {
   const { auth, user, deleteUser } = useAuth();
   const navigate = useNavigate();
+  const [urlPath, setUrlPath] = useState(window.location.pathname);
 
-  
+  useEffect(() => {
+    setUrlPath(window.location.pathname)
+  }, [urlPath])
+
+
   function logout() {
-    deleteUser()
-    navigate('/home')
+
+    Swal.fire({
+      title: 'Deseja realmente sair?',
+      width: '360',
+      color: '#545776',
+      icon:'question',
+      focusCancel: true,
+      showCancelButton: true,
+      confirmButtonText: 'Sim',
+      cancelButtonText: 'Não',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteUser()
+        navigate('/home')
+      } else {
+        return
+      }
+    })
   }
 
   function changeScreen(type) {
+
     if (type === 'login') {
       navigate('/login')
     }
     if (type === 'createUser') {
       navigate('/createUser')
     }
+    if (type === 'home') {
+      navigate('/home')
+    }
+
+    setTimeout(()=>{setUrlPath('')},1)
   }
 
 
+
+  //devemos trocar o 'user' pelo 'auth' na verificação para ver se tem alguém logado
   return (
     <section className="headerFull">
 
-      <img src="src/img/logo.png" alt="logotipo" />
+      <img onClick={()=>changeScreen('home')} className="logoImg" src="src/assets/logo1-3.svg" alt="logotipo" />
+
       <h2 className="logoTagline">Sinta-se em casa</h2>
 
-      <div className="asideHolder">
+      <div className="asideHolder" type={urlPath}>
 
         {user !== '' && (
           <div className="loggedIn">
@@ -45,11 +77,18 @@ export function Header() {
           </div>
         )}
 
-        {user=='' &&
+        {user == '' &&
           <>
-            <button className='btnHeader' onClick={() => changeScreen('createUser')}>Criar conta</button>
-            <button className='btnHeader' onClick={() => changeScreen('login')}>Iniciar sessão</button>
+            {(urlPath === '/login' || urlPath === '/home') &&
+              <button className='btnHeader' onClick={() => changeScreen('createUser')}>Criar conta</button>
+            }
+
+            {(urlPath === '/createUser' || urlPath === '/home') &&
+              <button className='btnHeader' onClick={() => changeScreen('login')}>Iniciar sessão</button>
+            }
+
           </>
+
         }
 
       </div>
