+++ front-end/airbabeach/src/Components/Sidebar.jsx
@@ -1,44 +1,60 @@
 import "./Sidebar.scss";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { FacebookLogo, LinkedinLogo, TwitterLogo, InstagramLogo, X } from 'phosphor-react'
 import { Link, useNavigate } from "react-router-dom";
 import { useAuth } from "../contexts/auth";
+import Swal from 'sweetalert2'
 
 
 export function Sidebar() {
   const { auth, user, deleteUser } = useAuth();
   const navigate = useNavigate();
 
+  const [urlPath, setUrlPath] = useState(window.location.pathname);
   const [sidebarDisplay, setSidebarDisplay] = useState(false);
   const [animation, setAnimation] = useState("");
 
-  const [isLogged, setLogged] = useState(false);
 
+  useEffect(() => {
+    setUrlPath(window.location.pathname)
+  }, [urlPath])
 
   function toggleSidebar() {
     if (!sidebarDisplay) {
       setSidebarDisplay(true);
       setAnimation("active")
-
     } else {
-      setAnimation("none");
       setSidebarDisplay(false)
-
+      setAnimation("none");
     }
+
+    setTimeout(()=>{setUrlPath('')},1)
   };
 
   function logout() {
-    deleteUser()
-    navigate('/home')
-    toggleSidebar()
-
+    Swal.fire({
+      title: 'Deseja realmente sair?',
+      width: '360',
+      color: '#545776',
+      icon:'question',
+      focusCancel: true,
+      showCancelButton: true,
+      confirmButtonText: 'Sim',
+      cancelButtonText: 'Não',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteUser()
+        navigate('/home')
+        toggleSidebar()
+      } else {
+        return
+      }
+    })
   }
 
 
-  //falta reconhecer a url atual para ativar os botões de redirecionamento
-  //falta criar o alert de confirmar deslogar ou não
-
-
+  //devemos trocar o 'user' pelo 'auth' na verificação para ver se tem alguém logado
   return (
     <section className="sideSection">
       <div className={`hamburger ${sidebarDisplay ? "active" : ""}`} onClick={toggleSidebar}>
@@ -66,19 +82,21 @@ export function Sidebar() {
 
           {!user &&
             <div className="loggedOut">
-              <h2 className="h2Menu">MENU</h2>
+              <h2 type={urlPath} className="h2Menu">MENU</h2>
               <ul className="menu">
 
-                <li className="menu-item"><Link to={'/createUser'} className="menu-link" onClick={toggleSidebar}>Criar conta</Link></li>
-                <hr className="hrStyle1" />
+                {(urlPath === '/login' || urlPath === '/home') && <li className="menu-item"><Link to={'/createUser'} className="menu-link" onClick={toggleSidebar}>Criar conta</Link></li>}
 
-                <li className="menu-item"><Link to={'/login'} className="menu-link" onClick={toggleSidebar}>Fazer login</Link></li>
+                {urlPath === '/home' && <hr className="hrStyle1" />}
+
+                {(urlPath === '/createUser' || urlPath === '/home') && <li className="menu-item"><Link to={'/login'} className="menu-link" onClick={toggleSidebar}>Fazer login</Link></li>}
+
               </ul>
             </div>
           }
 
           <div className="socialMidia">
-            {isLogged && <hr className="hrStyle2" />}
+            {user && <hr className="hrStyle2" />}
             <a href="#" className="socialMediaIcon" ><FacebookLogo size={28} weight="fill" /></a>
             <a href="#" className="socialMediaIcon" ><LinkedinLogo size={25} weight="bold" /></a>
             <a href="#" className="socialMediaIcon" ><TwitterLogo size={24} weight="fill" /></a>
